requirements.txt
openai
pinecone-client
langchain
python-dotenv
PyPDF2